requirements.txt
streamlit==1.38.0 
langchain==0.3.0 
langchain-openai==0.2.0 
langchain-community==0.3.0 
PyPDF2==3.0.1 
python-docx==1.1.2 
chromadb==0.5.5 
python-dotenv==1.0.1 
pyyaml==6.0.2